refactor(ui-web): extract select host component into separate component
So it's host component is cleaner
