+++ libs/ui-web/src/lib/components/components.ts
@@ -6,9 +6,10 @@ import { UiWebHeadingHostComponent } from './ui-web-heading-host';
 import { UiWebImageHostComponent } from './ui-web-image-host';
 import { UiWebInputHostComponent } from './ui-web-input-host';
 import { UiWebLabelHostComponent } from './ui-web-label-host';
+import { UiWebSelectComponent } from './ui-web-select';
 import { UiWebSelectHostComponent } from './ui-web-select-host';
 
-export const COMPONENTS: Type<any>[] = [];
+export const COMPONENTS: Type<any>[] = [UiWebSelectComponent];
 
 export const HOST_COMPONENTS: OrchestratorDynamicComponentType[] = [
   UiWebButtonHostComponent,
