+++ libs/ui-web/src/lib/components/index.ts
@@ -3,4 +3,5 @@ export * from './ui-web-label-host';
 export * from './ui-web-image-host';
 export * from './ui-web-heading-host';
 export * from './ui-web-input-host';
+export * from './ui-web-select';
 export * from './ui-web-select-host';
