+++ libs/ui-web/src/lib/components/ui-web-select-host/ui-web-select-config.ts
@@ -1,21 +1,7 @@
 import { Injectable } from '@angular/core';
 
 import { GenericFormAttributesConfig } from '../../types';
-
-export interface UiWebSelectOptionGroup {
-  children: (string | UiWebSelectOptionPair)[];
-  label?: string;
-  disabled?: boolean;
-}
-
-export interface UiWebSelectOptionPair {
-  label: string;
-  value: any;
-  disabled?: boolean;
-  selected?: boolean;
-}
-
-export type UiWebSelectOption = string | UiWebSelectOptionPair | UiWebSelectOptionGroup;
+import { UiWebSelectOption } from '../ui-web-select';
 
 @Injectable({ providedIn: 'root' })
 export class UiWebSelectConfig extends GenericFormAttributesConfig {
