+++ libs/ui-web/src/lib/components/ui-web-select-host/ui-web-select-host.component.html
@@ -1,4 +1,5 @@
-<select
+<orc-ui-web-select
+  [options]="config?.options"
   [name]="config?.name"
   [multiple]="config?.multiple"
   [size]="config?.size"
@@ -6,37 +7,4 @@
   [disabled]="config?.disabled"
   [tabIndex]="config?.tabindex"
   [autofocus]="config?.autofocus"
->
-  <ng-container
-    *ngFor="let option of config?.options"
-    [ngTemplateOutlet]="renderGroupOrOption"
-    [ngTemplateOutletContext]="{ $implicit: option }"
-  ></ng-container>
-</select>
-<ng-template #renderGroupOrOption let-groupOrOption>
-  <optgroup
-    *ngIf="isOptionGroup(groupOrOption); else: renderOpt"
-    [label]="groupOrOption.label"
-    [disabled]="groupOrOption.disabled"
-  >
-    <ng-container
-      *ngFor="let option of groupOrOption.children"
-      [ngTemplateOutlet]="renderOption"
-      [ngTemplateOutletContext]="{ $implicit: option }"
-    ></ng-container>
-  </optgroup>
-  <ng-template #renderOpt>
-    <ng-container *ngTemplateOutlet="renderOption; context: { $implicit: groupOrOption }"></ng-container>
-  </ng-template>
-</ng-template>
-<ng-template #renderOption let-option>
-  <option
-    *ngIf="isOptionPair(option); else: optionString"
-    [value]="option.value"
-    [disabled]="option.disabled"
-    [selected]="option.selected"
-  >{{ option.label }}</option>
-  <ng-template #optionString>
-    <option [value]="option">{{ option }}</option>
-  </ng-template>
-</ng-template>
+></orc-ui-web-select>
