+++ libs/ui-web/src/lib/components/ui-web-select-host/ui-web-select-host.component.ts
@@ -1,12 +1,7 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { DynamicComponent, OrchestratorDynamicComponent } from '@orchestrator/core';
 
-import {
-  UiWebSelectConfig,
-  UiWebSelectOption,
-  UiWebSelectOptionGroup,
-  UiWebSelectOptionPair,
-} from './ui-web-select-config';
+import { UiWebSelectConfig } from './ui-web-select-config';
 
 @Component({
   selector: 'orc-ui-web-select-host',
@@ -17,12 +12,4 @@ import {
 @DynamicComponent({ config: UiWebSelectConfig })
 export class UiWebSelectHostComponent implements OrchestratorDynamicComponent<UiWebSelectConfig> {
   @Input() config: UiWebSelectConfig;
-
-  isOptionGroup(option: UiWebSelectOption): option is UiWebSelectOptionGroup {
-    return typeof option === 'object' && 'children' in option;
-  }
-
-  isOptionPair(option: UiWebSelectOption): option is UiWebSelectOptionPair {
-    return typeof option === 'object' && 'value' in option;
-  }
 }
