+++ libs/ui-web/src/lib/components/ui-web-select/index.ts
@@ -1,0 +1,2 @@
+export * from './ui-web-select.component';
+export * from './types';
