+++ libs/ui-web/src/lib/components/ui-web-select/types.ts
@@ -1,0 +1,14 @@
+export interface UiWebSelectOptionGroup {
+  children: (string | UiWebSelectOptionPair)[];
+  label?: string;
+  disabled?: boolean;
+}
+
+export interface UiWebSelectOptionPair {
+  label: string;
+  value: any;
+  disabled?: boolean;
+  selected?: boolean;
+}
+
+export type UiWebSelectOption = string | UiWebSelectOptionPair | UiWebSelectOptionGroup;
